tasky/telegram_bot/requirements.txt
python-telegram-bot==20.6
httpx
psycopg2-binary

